+++ Server/Server/Gameplay/Formulas.cs
@@ -13,9 +13,9 @@ namespace Server.Gameplay
             return (int)stamina * 25;
         }
 
-        public static int XPToLevel(float xp)
+        public static byte XPToLevel(float xp)
         {
-            return (int)Math.Pow(xp + 9, 1.0 / 3) - 2;
+            return (byte)(Math.Pow(xp + 9, 1.0 / 3) - 2);
         }
 
         public static int LevelToXP(float level)
@@ -23,7 +23,7 @@ namespace Server.Gameplay
             return (int)Math.Pow(level + 2, 3) - 8;
         }
 
-        public static int LevelToPower(int level)
+        public static int LevelToPower(byte level)
         {
             return 100 + level * 20;
         }
